refactored a lot
